+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+
+## [0.3.0] - 2021-09-14
+**BREAKING CHANGES**:
+`MsgPack.default` has been removed and `MsgPack` can be used directly instead, e.g.:
+
+```kotlin
+MsgPack.encodeToByteArray(...)
+```
+instead of:
+
+```kotlin
+MsgPack.default.encodeToByteArray(...)
+```
+
 ### Added
 - Upgraded kotlin version to 1.5.0
 - Upgraded kotlinx-serialization version to 1.2.2
@@ -38,9 +52,10 @@ All notable changes to this project will be documented in this file. This change
 - `MsgPackDynamicSerializer` as placeholder for future [contextual serializer](https://github.com/Kotlin/kotlinx.serialization/blob/master/docs/serializers.md#contextual-serialization)
 - Full implementation of msgpack spec excluding extension types and bin format family
 
-[Unreleased]: https://github.com/esensar/kotlinx-serialization-msgpack/compare/0.2.1...main
+[Unreleased]: https://github.com/esensar/kotlinx-serialization-msgpack/compare/0.3.0...main
 [0.2.0]: https://github.com/esensar/kotlinx-serialization-msgpack/compare/0.1.0...0.2.0
 [0.2.1]: https://github.com/esensar/kotlinx-serialization-msgpack/compare/0.2.0...0.2.1
+[0.3.0]: https://github.com/esensar/kotlinx-serialization-msgpack/compare/0.2.1...0.3.0
 [i6]: https://github.com/esensar/kotlinx-serialization-msgpack/issues/6
 [i9]: https://github.com/esensar/kotlinx-serialization-msgpack/issues/9
 [i10]: https://github.com/esensar/kotlinx-serialization-msgpack/issues/10